+++ tests/test_HOAS.v
@@ -230,11 +230,14 @@ Elpi Query lp:{{
 Elpi Command primitive.
 Elpi Accumulate lp:{{
 main [trm T] :-
-  std.assert! (coq.reduction.vm.whd_all T _ T1) "normal form is not an opinion",
-  std.assert! (coq.reduction.cbv.whd_all T T1) "normal form is not an opinion",
+  std.assert! (coq.reduction.native.norm T _ T1) "normal form is not an opinion",
+  std.assert! (coq.reduction.vm.norm T _ T1) "normal form is not an opinion",
+  std.assert! (coq.reduction.cbv.norm T T1) "normal form is not an opinion",
+  std.assert! (coq.reduction.lazy.norm T T1) "normal form is not an opinion",
   std.assert! (coq.reduction.lazy.whd_all T T1) "normal form is not an opinion",
   coq.say "Raw term:" T "\nNice term:" {coq.term->string T} "\nRed:" {coq.term->string T1}.
 }}.
+Elpi Typecheck.
 From Coq Require Import Int63.
 Open Scope int63_scope.
 Elpi primitive (2000000003333002 + 1).